% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fmt.R
\name{compute_positions}
\alias{compute_positions}
\title{Compute Start and End Positions}
\usage{
compute_positions(ptbl_fmt)
}
\arguments{
\item{ptbl_fmt}{format tibble for which start and end positions must be computed}
}
\value{
tbl_fmt format tibble with new start and end positions
}
\description{
Start and end positions are re-computed based on the width column given in
the format tibble ptbl_fmt. Any existing start and end positions are updated.
}
