% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_fmt.R
\name{delete_row}
\alias{delete_row}
\title{Delete Row from Format Tibble}
\usage{
delete_row(ptbl_imp_fmt, pvec_del_row_idx)
}
\arguments{
\item{ptbl_imp_fmt}{given format tibble}

\item{pvec_del_row_idx}{row indices for rows to be deleted}
}
\value{
tbl_imp_fmt format tibble without rows to be delted
}
\description{
Given a tibble with format information, a given number of rows
are deleted. After the deletion, the start and the end positions
must be re-computed
}
