% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fmt.R
\name{extract_gal_fmt}
\alias{extract_gal_fmt}
\title{Extract GAL Export Format}
\usage{
extract_gal_fmt(
  ps_sql_prg_path,
  ps_sql_exp_routine = "ExportGAL",
  ps_exp_pattern = "PA_EXP.sFormat"
)
}
\arguments{
\item{ps_sql_prg_path}{path to sql program}

\item{ps_sql_exp_routine}{name of export routine}

\item{ps_exp_pattern}{pattern to search for export commands}
}
\value{
tbl_gal_fmt tibble with the column headers and the column width
}
\description{
Extraction of format specification of exported GAL data based on SQL export
routine. This function is a wrapper which uses the generic export function
\code{extract_generic_fmt}.
}
\details{
Usefulness of this function depends on the format of the SQL export routine.
If that format changes, the result might be completely unusable. This function
should only be used once to extract the format. Any updates should be done
on already existing fmt-input files.
}
\examples{
\dontrun{
s_sql_prg_path <- system.file("extdata", "zws_gal_pgb.sql", package = "qfmtchk")
extract_gal_fmt(ps_sql_prg_path = s_sql_prg_path)
ps_sql_prg_path="/Users/pvr/Data/Projects/argus_sql/trunk/Argus/zws_gal_pgb.sql"
ps_sql_exp_routine="ExportGAL"
ps_exp_pattern = "PA_EXP.sFormat"
}

}
