% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fmt.R
\name{extract_generic_fmt}
\alias{extract_generic_fmt}
\title{Generic Extraction of Data Format Information}
\usage{
extract_generic_fmt(ps_sql_prg_path, ps_sql_exp_routine, ps_exp_pattern)
}
\arguments{
\item{ps_sql_prg_path}{path to sql program}

\item{ps_sql_exp_routine}{name of export routine}
}
\value{
tbl_fmt tibble with data column name and the width of the column
}
\description{
The format information is extracted from the SQL-Export routine. This
routine has to be stored in a textfile whose path is given by the
parameter ps_sql_prg_path. In that SQL program all code between the
beginning and the end of the specified routine name (ps_sql_exp_routine)
is investigated for the pattern given by ps_exp_pattern.
}
\details{
The format extraction works only, if the export routine is written in
a specific way. If these assumptions are not met, the extraction does
not work.
}
