% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fmt.R
\name{extract_kldat_fmt}
\alias{extract_kldat_fmt}
\title{Extract FMT Information For KLDAT}
\usage{
extract_kldat_fmt(
  ps_sql_prg_path,
  ps_sql_exp_routine = "ImportSchlachtdatenTvd",
  ps_exp_pattern = "rowSTV.STV"
)
}
\arguments{
\item{ps_sql_prg_path}{path to SQL program from where format is extracted}

\item{ps_sql_exp_routine}{name of export routine in SQL program}

\item{ps_exp_pattern}{pattern used to find export lines}
}
\value{
tbl_kldat_fmt tibble with extracted format
}
\description{
The format specification used in the SQL program that imports the carcass
data into the database are extracted and collected into a tibble where
each column in the KLDAT input file is listed with the start position.
}
