% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_fmt.R
\name{improve_fmt}
\alias{improve_fmt}
\title{Improve Format Specification}
\usage{
improve_fmt(
  ptbl_fmt,
  pvec_del_row_idx = NULL,
  ptbl_col_rename = NULL,
  ptbl_fmt_add = NULL
)
}
\arguments{
\item{ptbl_fmt}{original format tibble which needs to be improved}

\item{pvec_del_row_idx}{vector of row indices with rows to be deleted}

\item{ptbl_col_rename}{tibble with row indices and new column names}

\item{ptbl_fmt_add}{tibble with fmt columns to be added}
}
\value{
tbl_imp_fmt improved format tibble
}
\description{
The improvement consists of removing rows from the fmt tibble, changing
the name of columns or of adding more rows to the fmt tibble.
}
\details{
\subsection{title: Improvement Functions for Format Specification}{

The vector of row indices for the rows to be deleted can be specified with
a simple numberic vector. The list of column names to be renamed can be
specified with a tibble (ptbl_col_rename) with a column RowIndex and a
second column NewName where the first gives the row index where the column
name should be renamed and the second column contains the new column
name.
}
}
\examples{
\dontrun{
improve_fmt(ptbl_fmt = tbl_gal_fmt, pvec_del_row_idx = c(4:6))
improve_fmt(ptbl_fmt = tbl_gal_fmt,
            pvec_del_row_idx = 4:6,
            ptbl_col_rename = tibble::tibble(RowIndex = c(3,6,9),
                                             NewName=c("BlutanteilAnimal",
                                                       "BlutanteilVater",
                                                       "Traechtigkeitsdauer")))
}

}
