% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql_prg.R
\name{read_sql_prg}
\alias{read_sql_prg}
\title{Read SQL Program Code}
\usage{
read_sql_prg(ps_sql_prg_path, ps_sql_exp_routine, ps_exp_pattern)
}
\arguments{
\item{ps_sql_prg_path}{path to sql program to read}

\item{ps_sql_exp_routine}{name of the routine from where format should be extracted}

\item{ps_exp_pattern}{pattern which indicate format information}
}
\value{
vec_exp_prg vector with sql statements that contain the format
}
\description{
The SQL code of the data export program is read and the part with the
export statements is returned. This part is expected to contain the
format information.
}
