% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_fmt.R
\name{rename_variable}
\alias{rename_variable}
\title{Rename Format Check Variables}
\usage{
rename_variable(ptbl_imp_fmt, ptbl_col_rename)
}
\arguments{
\item{ptbl_imp_fmt}{format tibble with old variable names}

\item{ptbl_col_rename}{tibble containing indices and new names}
}
\value{
tbl_imp_fmt format tibble with new names
}
\description{
For a given format tibble certain column variable names can be re-named.
}
