% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fmt.R
\name{output_fmt}
\alias{output_fmt}
\title{Write FMT Info To File}
\usage{
output_fmt(ps_fmt_outfile, ptbl_fmt)
}
\arguments{
\item{ps_fmt_outfile}{output file for fmt information}

\item{ptbl_fmt}{tibble containing fmt information}
}
\description{
The tibble with the fmt information consisting of column names and
positions is taken and converted into an input file for the chckfwf
program.
}
\examples{
\dontrun{
output_fmt(ps_fmt_outfile = "gal_check.fmt", ptbl_fmt = tbl_gal_fmt)
}

}
